# Please enter a commit message to explain why this merge is necessary,
# especially if it merges an updated upstream into a topic branch.
#
# Lines starting with '#' will be ignored, and an empty message aborts
# the commit.

+++ src/stan/mcmc/chains.hpp
@@ -319,7 +319,7 @@ namespace stan {
         : param_names_(param_names) { }
 
       chains(const std::vector<std::string>& param_names) : param_names_(param_names.size()) {
-          for (int i = 0; i < param_names.size(); i++)
+          for (size_t i = 0; i < param_names.size(); i++)
               param_names_(i) = param_names[i];
       }
